% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_A.R
\name{calc_cofactor}
\alias{calc_cofactor}
\title{Function to calculate a cofactor matrix}
\usage{
calc_cofactor(data)
}
\arguments{
\item{data}{Matrix with n rows (samples) and p columns (variables). All data 
should be continuous values and should not have missing values.}
}
\value{
\code{calc_cofactor} returns a cofactor matrix of size p x p.
}
\description{
\code{calc_cofactor} calculates a cofactor matrix.
}
\examples{
# 40 data for versicolor in the iris dataset                            
iris_versicolor <- iris[61:100, -5] 
                            
calc_cofactor(cov(iris_versicolor))
 
}
\seealso{
\code{\link{MTA}}
}

